Begin breaking up parsers into child modules
